#34 fix: Content Response 에 조회수 추가 및 조회에서 배치 타임 기반으로 변경

+++ src/main/java/com/example/server/domain/quiz/dto/QuizChoiceResponse.java
@@ -1,19 +1,20 @@
 package com.example.server.domain.quiz.dto;
 
 import com.example.server.domain.quiz.entity.QuizChoice;
+
 import lombok.Builder;
 
 @Builder
 public record QuizChoiceResponse(
-        Integer quizChoiceId,
-        Integer choiceNo,
-        String choiceText
+	Long quizChoiceId,
+	Integer choiceNo,
+	String choiceText
 ) {
-    public static QuizChoiceResponse from(QuizChoice c) {
-        return QuizChoiceResponse.builder()
-                .quizChoiceId(c.getQuizChoiceId())
-                .choiceNo(c.getChoiceNo())
-                .choiceText(c.getChoiceText())
-                .build();
-    }
+	public static QuizChoiceResponse from(QuizChoice c) {
+		return QuizChoiceResponse.builder()
+			.quizChoiceId(c.getQuizChoiceId())
+			.choiceNo(c.getChoiceNo())
+			.choiceText(c.getChoiceText())
+			.build();
+	}
 }
